requirements.txt
asyncpg
discord
DateTime
asyncio
statcord.py
dblpy